Edit button in admin panel and loading photos in database
I've added an edit button in the admin panel where you can edit a product's details. I've also linked the database with cloudinary and multer. With multer you can add the photos from the admin panel. When you click on add product/update product if editing, the photos is being sent to a folder in my cloudinary cloud, where its being stored safely. Last thing to fix is the orderCount.